Add SubscriptionPagesUseSubscriptionFeature Free Trial tests. Cleanup more experiment-related code.
