+++ iOS/DuckDuckGo/SubscriptionDebugViewController.swift
@@ -57,7 +57,6 @@ final class SubscriptionDebugViewController: UITableViewController {
         Sections.pixels: "Promo Pixel Parameters",
         Sections.metadata: "StoreKit Metadata",
         Sections.regionOverride: "Region override for App Store Sandbox",
-        Sections.featureFlags: "Feature Flags"
     ]
 
     enum Sections: Int, CaseIterable {
@@ -69,7 +68,6 @@ final class SubscriptionDebugViewController: UITableViewController {
         case pixels
         case metadata
         case regionOverride
-        case featureFlags
     }
 
     enum AuthorizationRows: Int, CaseIterable {
@@ -111,13 +109,8 @@ final class SubscriptionDebugViewController: UITableViewController {
         case currentRegionOverride
     }
 
-    enum FeatureFlagRows: Int, CaseIterable {
-        case privacyProFreeTrialJan25
-    }
-
     private var storefrontID = "Loading"
     private var storefrontCountryCode = "Loading"
-    private let freeTrialKey = FreeTrialsFeatureFlagExperiment.Constants.featureFlagOverrideKey
 
     override func numberOfSections(in tableView: UITableView) -> Int {
         return Sections.allCases.count
@@ -277,15 +270,6 @@ final class SubscriptionDebugViewController: UITableViewController {
                 break
             }
 
-        case .featureFlags:
-            switch FeatureFlagRows(rawValue: indexPath.row) {
-            case .privacyProFreeTrialJan25:
-                cell.textLabel?.text = "privacyProFreeTrialJan25"
-                cell.accessoryType = UserDefaults.standard.bool(forKey: freeTrialKey) ? .checkmark : .none
-            case .none:
-                break
-            }
-
         case .none:
             break
         }
@@ -303,7 +287,6 @@ final class SubscriptionDebugViewController: UITableViewController {
         case .pixels: return PixelsRows.allCases.count
         case .metadata: return MetadataRows.allCases.count
         case .regionOverride: return RegionOverrideRows.allCases.count
-        case .featureFlags: return FeatureFlagRows.allCases.count
         case .none: return 0
         }
     }
@@ -346,11 +329,6 @@ final class SubscriptionDebugViewController: UITableViewController {
             break
         case .regionOverride:
             break
-        case .featureFlags:
-            switch FeatureFlagRows(rawValue: indexPath.row) {
-            case .privacyProFreeTrialJan25: togglePrivacyProFreeTrialJan25Flag()
-            default: break
-            }
         case .none:
             break
         }
@@ -497,12 +475,6 @@ final class SubscriptionDebugViewController: UITableViewController {
                 Randomized: \(reportedParameters.joined(separator: ", "))
                 """
         showAlert(title: "", message: message)
-    }
-
-    private func togglePrivacyProFreeTrialJan25Flag() {
-        let currentValue = UserDefaults.standard.bool(forKey: freeTrialKey)
-        UserDefaults.standard.set(!currentValue, forKey: freeTrialKey)
-        tableView.reloadData()
     }
 
     private func syncAppleIDAccount() {
